Comments and minor fixes
