+++ tests/Microsoft.Identity.Test.Unit/pop/PoPTests.cs
@@ -398,7 +398,8 @@ namespace Microsoft.Identity.Test.Unit.Pop
                 mockBroker.IsPopSupported.Returns(true);
                 mockBroker.AcquireTokenSilentAsync(
                     Arg.Any<AuthenticationRequestParameters>(),
-                    Arg.Any<AcquireTokenSilentParameters>()).Returns(CreateMsalRunTimeBrokerTokenResponse(null, Constants.PoPAuthHeaderPrefix));
+                    Arg.Any<AcquireTokenSilentParameters>()).Returns(
+                        MockHelpers.CreateMsalRunTimeBrokerTokenResponse(null, Constants.PoPAuthHeaderPrefix));
 
                 var pcaBuilder = PublicClientApplicationBuilder.Create(TestConstants.ClientId)
                     .WithTestBroker(mockBroker)
@@ -576,6 +577,8 @@ namespace Microsoft.Identity.Test.Unit.Pop
 
             Assert.IsTrue(JObject.DeepEquals(jwkInConfig, jwkInToken));
         }
+
+       
     }
 }
